fix apply response schema when needed
