+++ lib/reply.js
@@ -3,7 +3,9 @@
 const eos = require('readable-stream').finished
 const statusCodes = require('http').STATUS_CODES
 const flatstr = require('flatstr')
+const FJS = require('fast-json-stringify')
 const {
+  kSchemaResponse,
   kFourOhFourContext,
   kReplyErrorHandlerCalled,
   kReplySent,
@@ -20,12 +22,21 @@ const {
 const { hookRunner, hookIterator, onSendHookRunner } = require('./hooks')
 const validation = require('./validation')
 const serialize = validation.serialize
-const serializeError = validation.serializeError
 
 const internals = require('./handleRequest')[Symbol.for('internals')]
 const loggerUtils = require('./logger')
 const now = loggerUtils.now
 const wrapThenable = require('./wrapThenable')
+
+const serializeError = FJS({
+  type: 'object',
+  properties: {
+    statusCode: { type: 'number' },
+    code: { type: 'string' },
+    error: { type: 'string' },
+    message: { type: 'string' }
+  }
+})
 
 const CONTENT_TYPE = {
   JSON: 'application/json; charset=utf-8',
@@ -522,13 +533,20 @@ function handleError (reply, error, cb) {
     return
   }
 
-  // TODO shallow copy bad idea in hot path
-  const payload = serializeError(reply.context, {
-    error: statusCodes[statusCode + ''],
-    message: error.message || '',
-    ...error,
-    statusCode
-  }, statusCode)
+  const serializerFn = getSchemaSerializer(reply.context, statusCode)
+  const payload = (serializerFn === false)
+    ? serializeError({
+      error: statusCodes[statusCode + ''],
+      code: error.code,
+      message: error.message || '',
+      statusCode: statusCode
+    })
+    : serializerFn(Object.create(error, {
+      error: { value: statusCodes[statusCode + ''] },
+      message: { value: error.message || '' },
+      statusCode: { value: statusCode }
+    }))
+
   flatstr(payload)
   reply[kReplyHeaders]['content-type'] = CONTENT_TYPE.JSON
   reply[kReplyHeaders]['content-length'] = '' + Buffer.byteLength(payload)
@@ -636,6 +654,22 @@ function notFound (reply) {
   } else {
     internals.preHandlerCallback(null, reply.request, reply)
   }
+}
+
+// TODO duplicated code from validation.js
+function getSchemaSerializer (context, statusCode) {
+  const responseSchemaDef = context[kSchemaResponse]
+  if (!responseSchemaDef) {
+    return false
+  }
+  if (responseSchemaDef[statusCode]) {
+    return responseSchemaDef[statusCode]
+  }
+  const fallbackStatusCode = (statusCode + '')[0] + 'xx'
+  if (responseSchemaDef[fallbackStatusCode]) {
+    return responseSchemaDef[fallbackStatusCode]
+  }
+  return false
 }
 
 function noop () { }
