+++ lib/validation.js
@@ -1,7 +1,5 @@
 'use strict'
 
-const FJS = require('fast-json-stringify')
-
 const {
   kSchemaHeaders: headersSchema,
   kSchemaParams: paramsSchema,
@@ -9,16 +7,6 @@ const {
   kSchemaBody: bodySchema,
   kSchemaResponse: responseSchema
 } = require('./symbols')
-
-const fastSerializeError = FJS({
-  type: 'object',
-  properties: {
-    statusCode: { type: 'number' },
-    code: { type: 'string' },
-    error: { type: 'string' },
-    message: { type: 'string' }
-  }
-})
 
 function compileSchemasForSerialization (context, compile) {
   if (!context.schema || !context.schema.response) {
@@ -135,30 +123,10 @@ function serialize (context, data, statusCode) {
   return JSON.stringify(data)
 }
 
-/**
- * This function is called whenever an instance of `Error`
- * must be serialized
- */
-function serializeError (context, data, statusCode) {
-  const responseSchemaDef = context[responseSchema]
-  if (!responseSchemaDef) {
-    return fastSerializeError(data)
-  }
-  if (responseSchemaDef[statusCode]) {
-    return responseSchemaDef[statusCode](data)
-  }
-  const fallbackStatusCode = (statusCode + '')[0] + 'xx'
-  if (responseSchemaDef[fallbackStatusCode]) {
-    return responseSchemaDef[fallbackStatusCode](data)
-  }
-  return fastSerializeError(data)
-}
-
 module.exports = {
   symbols: { bodySchema, querystringSchema, responseSchema, paramsSchema, headersSchema },
   compileSchemasForValidation,
   compileSchemasForSerialization,
   validate,
-  serialize,
-  serializeError
+  serialize
 }
