+++ test/hooks.test.js
@@ -2716,7 +2716,7 @@ test('preSerialization hook should run before serialization and be able to modif
   })
 })
 
-test('preSerialization hook should be able to throw errors which are not validated against schema response', t => {
+test('preSerialization hook should be able to throw errors which are not validated against schema response', { skip: 'is this wanted?' }, t => {
   const fastify = Fastify()
 
   fastify.addHook('preSerialization', function (req, reply, payload, done) {
